# docs/swagger.yaml
definitions:
  gorm_model.AuthMeResp:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      profile_picture:
        type: string
      system_role_name:
        type: string
    type: object
  gorm_model.JobRoleResp:
    properties:
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      sector:
        $ref: '#/definitions/gorm_model.SectorResp'
      sector_id:
        type: string
    type: object
  gorm_model.ProvinsiResp:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  gorm_model.SectorResp:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  gorm_model.UserResp:
    properties:
      account_status:
        type: string
      assigned_role_id:
        type: string
      bidang:
        type: string
      birthdate:
        type: string
      candidate_level:
        type: string
      email:
        type: string
      gpa:
        type: number
      id:
        type: string
      is_bookmark:
        type: boolean
      job_roles:
        items:
          $ref: '#/definitions/gorm_model.JobRoleResp'
        type: array
      job_title_id:
        type: string
      kabupaten_kota_id:
        type: string
      major:
        type: string
      must_reset_password:
        type: boolean
      name:
        type: string
      phone_number:
        type: string
      portofolio_link:
        type: string
      profile_picture:
        type: string
      province:
        $ref: '#/definitions/gorm_model.ProvinsiResp'
      recruitment_status_hex_code:
        type: string
      recruitment_status_id:
        type: string
      recruitment_status_name:
        type: string
      school_university:
        type: string
      summary:
        type: string
      system_role_name:
        type: string
      tech_stack:
        type: string
      unavailable_until:
        type: string
      years_experience:
        type: integer
    type: object
...
  request_model.CreateRecruitmentStatusRequest:
    properties:
      hex_code:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    required:
    - name
    type: object
...
  request_model.PatchInterviewStatusRequest:
    properties:
      interview_id:
        type: string
      status:
        type: string
    required:
    - interview_id
    - status
    type: object
  request_model.PatchUserRecruitmentStatusRequest:
    properties:
      candidate_level:
        type: string
      recruitment_status_id:
        type: string
    type: object
...
  request_model.UpdatePasswordRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
  request_model.UpdateProfileRequest:
    properties:
      birthdate:
        description: Expecting YYYY-MM-DD
        type: string
      gpa:
        type: number
      job_role_ids:
        items:
          type: string
        type: array
      kabupaten_kota_id:
        type: string
      major:
        type: string
      name:
        type: string
      phone_number:
        type: string
      portofolio_link:
        type: string
      school_university:
        type: string
      summary:
        type: string
      tech_stack:
        items:
          type: string
        type: array
      years_experience:
        type: integer
    type: object
  request_model.UpdateRecruitmentStatusRequest:
    properties:
      hex_code:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    required:
    - name
    type: object
...
  response.Base:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
      validation:
        additionalProperties:
          type: string
        type: object
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /auth/me:
    get:
      consumes:
      - application/json
      description: Get detail current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gorm_model.AuthMeResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Detail user
      tags:
      - Auth
...
  /interview/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Patch Status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_model.PatchInterviewStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Base'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Patch Interview Status
      tags:
      - Interview
...
  /profile:
    get:
      consumes:
      - application/json
      description: Get detailed profile of current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gorm_model.UserResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Get Profile
      tags:
      - Profile
    put:
      consumes:
      - application/json
      description: Update the profile of the current user
      parameters:
      - description: Profile update data
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request_model.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Update Profile
      tags:
      - Profile
  /profile/password:
    put:
      consumes:
      - application/json
      description: Change the current user's password. Mandatory if must_reset_password
        is true.
      parameters:
      - description: Password update data
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request_model.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Update Password
      tags:
      - Profile
...
  /recruitment-statuses:
    get:
      consumes:
      - application/json
      description: Get All Recruitment Status
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit per page
        in: query
        name: limit
        type: integer
      - description: Recruitment Status ID
        in: query
        name: recruitment_status_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Get All Recruitment Status
      tags:
      - Recruitment Status
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_model.CreateRecruitmentStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Create Recruitment Status
      tags:
      - Recruitment Status
  /recruitment-statuses/{id}:
    delete:
      parameters:
      - description: Recruitment Status ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Delete Recruitment Status
      tags:
      - Recruitment Status
    get:
      consumes:
      - application/json
      description: Get Recruitment Status By ID
      parameters:
      - description: Recruitment Status ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Get Recruitment Status By ID
      tags:
      - Recruitment Status
    put:
      consumes:
      - application/json
      parameters:
      - description: Recruitment Status ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_model.UpdateRecruitmentStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Update Recruitment Status
      tags:
      - Recruitment Status
...
  /users/{id}/recruitment-status:
    patch:
      consumes:
      - application/json
      description: Update a user's recruitment status and candidate level by superadmin
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Recruitment status patch object
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request_model.PatchUserRecruitmentStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Base'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Patch User Recruitment Status and Candidate Level
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: 'Insert JWT Token with format: **Bearer {your token}**'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"